# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: openleash API
  description: |
    Local-first authorization and cryptographic proof sidecar for AI agents.

    OpenLeash evaluates YAML-based policies against agent action requests and,
    when an action is allowed, issues PASETO v4.public proof tokens that
    counterparties can verify offline or online.

    ## Authentication

    The API uses four security zones:

    - **Public** — No authentication. Health checks, public key distribution,
      proof verification, and the playground.
    - **Agent** — Ed25519 request signing via five headers (`X-Agent-Id`,
      `X-Timestamp`, `X-Nonce`, `X-Body-Sha256`, `X-Signature`). Used for
      `/v1/authorize` and `/v1/agent/*` endpoints.
    - **Owner** — PASETO v4.public session token via `Authorization: Bearer`
      header (or `openleash_session` cookie for GUI). Issued on login.
      Used for all `/v1/owner/*` endpoints.
    - **Admin** — Bearer token and/or localhost check depending on the
      `admin.mode` setting (`localhost`, `token`, or `localhost_or_token`).
      Used for all `/v1/admin/*` endpoints.

    ## Typical flow

    1. An admin creates an **owner** via the Admin API and generates a
       **setup invite**.
    2. The owner completes setup (`POST /v1/owner/setup`) with the invite
       token and a passphrase, then logs in (`POST /v1/owner/login`).
    3. The owner creates **policies** via the Owner API.
    4. An agent calls `/v1/agents/registration-challenge` then
...
servers:
...
paths:
...
  /v1/admin/owners:
    get:
      summary: List all owners
      operationId: listOwners
      tags: [Admin]
      description: |
        Returns all owner principals in the system.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of owners.
          content:
            application/json:
              schema:
                type: object
                required: [owners]
                properties:
                  owners:
                    type: array
                    items:
                      $ref: "#/components/schemas/Owner"
    post:
      summary: Create a new owner
      operationId: createOwner
      tags: [Admin]
      description: |
        Creates a new owner principal. After creation, use
        `POST /v1/admin/owners/{ownerId}/setup-invite` to generate a
        setup invite for the owner.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [principal_type, display_name]
              properties:
                principal_type:
                  type: string
                  enum: [HUMAN, ORG]
                display_name:
                  type: string
                contact_identities:
                  type: array
                  items:
                    $ref: "#/components/schemas/ContactIdentity"
                government_ids:
                  type: array
                  items:
                    $ref: "#/components/schemas/GovernmentId"
                company_ids:
                  type: array
                  items:
                    $ref: "#/components/schemas/CompanyId"
                signatories:
                  type: array
                  items:
                    $ref: "#/components/schemas/Signatory"
                signatory_rules:
                  type: array
                  items:
                    $ref: "#/components/schemas/SignatoryRule"
      responses:
        "200":
          description: Owner created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Owner"
  /v1/admin/owners/{ownerId}/setup-invite:
    post:
      summary: Generate setup invite
      operationId: createSetupInvite
      tags: [Admin]
      description: |
        Generates a one-time setup invite for an owner. The plaintext
        invite token is returned once and cannot be retrieved again.
        The owner uses this token with `POST /v1/owner/setup` to set
        their passphrase.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OwnerId"
      responses:
        "200":
          description: Invite created.
          content:
            application/json:
              schema:
                type: object
                required: [invite_id, invite_token, expires_at]
                properties:
                  invite_id:
                    type: string
                    format: uuid
                  invite_token:
                    type: string
                    description: Plaintext invite token (one-time, give to owner securely).
                  expires_at:
                    type: string
                    format: date-time
  /v1/admin/owners/{ownerId}:
    delete:
      summary: Delete owner
      operationId: deleteOwner
      tags: [Admin]
      description: |
        Removes an owner from the system state. Does not delete
        associated agents or policies.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OwnerId"
      responses:
        "200":
          description: Owner deleted.
          content:
            application/json:
              schema:
                type: object
                required: [owner_principal_id, status]
                properties:
                  owner_principal_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: deleted
  /v1/admin/agents:
    get:
      summary: List all agents
      operationId: listAgents
      tags: [Admin]
      description: Returns all registered agents across all owners.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of agents.
          content:
            application/json:
              schema:
                type: object
                required: [agents]
                properties:
                  agents:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentRegistration"
  /v1/admin/policies:
    get:
      summary: List all policies
      operationId: listPolicies
      tags: [Admin]
      description: Returns all policies across all owners (read-only view).
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of policies.
          content:
            application/json:
              schema:
                type: object
                required: [policies]
                properties:
                  policies:
                    type: array
                    items:
                      $ref: "#/components/schemas/PolicyEntry"
  /v1/admin/policies/{policyId}:
    get:
      summary: Get policy details
      operationId: getPolicy
      tags: [Admin]
      description: Returns a specific policy with its YAML content (read-only).
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PolicyId"
      responses:
        "200":
          description: Policy with YAML.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyEntry"
        "404":
          description: Policy not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/admin/audit:
    get:
      summary: Full audit log
      operationId: getAuditLog
      tags: [Admin]
      description: Returns the complete audit log (unfiltered).
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: cursor
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuditEvent"
                  next_cursor:
                    type: integer
                    nullable: true
  /v1/admin/config:
    get:
      summary: Get server configuration
      operationId: getConfig
      tags: [Admin]
      description: |
        Returns the merged server configuration (defaults + config.yaml).
        Sensitive values (admin token) are redacted.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Server configuration.
          content:
            application/json:
              schema:
                type: object
                description: Full configuration object (structure varies).
  /v1/admin/state:
    get:
      summary: Get full state
      operationId: getState
      tags: [Admin]
      description: |
        Returns the full system state including owners, agents, policies,
        bindings, and server keys. This is the authoritative index.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Full state data.
          content:
            application/json:
              schema:
                type: object
                description: Complete StateData object.

  # ── Playground ──────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Admin token configured in `config.yaml` under `admin.token`.
        Required when `admin.mode` is `token` or for remote access in
        `localhost_or_token` mode.

    OwnerSession:
      type: http
      scheme: bearer
      description: |
        PASETO v4.public session token issued by `POST /v1/owner/login`.
        Contains `purpose: 'owner_session'` claim to distinguish from
        proof tokens. Default TTL is 8 hours (configurable via
        `sessions.ttl_seconds` in `config.yaml`).
  parameters:
    OwnerId:
      name: ownerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the owner principal.
    PolicyId:
      name: policyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: UUID of the policy.
...
  schemas:
    Error:
      type: object
      description: Standard error envelope returned by all endpoints on failure.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code (e.g. `INVALID_REQUEST`, `AGENT_NOT_FOUND`).
            message:
              type: string
              description: Human-readable error message.
            details:
              type: object
              description: Additional structured details (e.g. Zod validation errors).
...
    AgentRegistration:
      type: object
      description: A registered AI agent.
      required: [agent_principal_id, agent_id, owner_principal_id, status, created_at]
      properties:
        agent_principal_id:
          type: string
          format: uuid
          description: Server-assigned UUID for this agent.
        agent_id:
          type: string
          description: The agent's self-chosen identifier.
        owner_principal_id:
          type: string
          format: uuid
          description: UUID of the owner this agent belongs to.
        status:
          type: string
          description: Agent status (e.g. `ACTIVE`, `REVOKED`).
        created_at:
          type: string
          format: date-time
    Owner:
      type: object
      description: An owner principal (human or organization) that holds policies and owns agents.
      required: [owner_principal_id, principal_type, display_name, status, created_at]
      properties:
        owner_principal_id:
          type: string
          format: uuid
        principal_type:
          type: string
          enum: [HUMAN, ORG]
        display_name:
          type: string
        status:
          type: string
        identity_assurance_level:
          type: integer
          description: Computed assurance level based on verified identity data (0-3).
        created_at:
          type: string
          format: date-time
    ContactIdentity:
      type: object
      description: A contact method (email, phone, etc.) attached to an owner.
      properties:
        contact_id:
          type: string
          format: uuid
          description: Auto-generated if not provided.
        type:
          type: string
          enum: [EMAIL, PHONE, WEBSITE]
        value:
          type: string
        verified:
          type: boolean
        added_at:
          type: string
          format: date-time
    GovernmentId:
      type: object
      description: A government-issued ID (passport, national ID, etc.) for a HUMAN owner.
      properties:
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code.
        id_type:
          type: string
          description: Type of government ID (e.g. `PERSONNUMMER`, `SSN`, `PASSPORT`).
        id_value:
          type: string
        verification_status:
          type: string
          enum: [UNVERIFIED, FORMAT_VALID, VERIFIED]
          description: Automatically upgraded to `FORMAT_VALID` if the format passes validation.
        added_at:
          type: string
          format: date-time
    CompanyId:
      type: object
      description: A company registration number (LEI, DUNS, etc.) for an ORG owner.
      properties:
        id_type:
          type: string
          description: Type of company ID (e.g. `LEI`, `DUNS`, `SE_ORGNR`).
        id_value:
          type: string
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code.
        verification_status:
          type: string
          enum: [UNVERIFIED, FORMAT_VALID, VERIFIED]
        added_at:
          type: string
          format: date-time
    Signatory:
      type: object
      description: An authorized signatory for an ORG owner, referencing a HUMAN owner.
      properties:
        signatory_id:
          type: string
          format: uuid
          description: Auto-generated if not provided.
        human_principal_id:
          type: string
          format: uuid
          description: UUID of the HUMAN owner principal who is the signatory.
        role:
          type: string
          description: Role of the signatory (e.g. `CEO`, `CFO`, `DIRECTOR`).
        added_at:
          type: string
          format: date-time
    SignatoryRule:
      type: object
      description: A quorum or approval rule governing when an ORG owner's actions are authorized.
      properties:
        rule_id:
          type: string
          format: uuid
          description: Auto-generated if not provided.
        rule_type:
          type: string
          description: Type of rule (e.g. `QUORUM`).
        min_signatories:
          type: integer
          description: Minimum number of signatories required.
        required_roles:
          type: array
          items:
            type: string
          description: Roles that must be represented among the signatories.
    PolicyEntry:
      type: object
      description: A stored authorization policy with its binding information.
      required: [policy_id, owner_principal_id, path]
      properties:
        policy_id:
          type: string
          format: uuid
        owner_principal_id:
          type: string
          format: uuid
        applies_to_agent_principal_id:
          type: string
          format: uuid
          nullable: true
          description: If set, this policy only applies to this specific agent.
        path:
          type: string
          description: File path of the stored policy YAML relative to the data directory.
        policy_yaml:
          type: string
          description: The full YAML content of the policy (present on GET, absent on list errors).
        error:
          type: string
          description: Present instead of `policy_yaml` if the file could not be read.
...
    AuditEvent:
      type: object
      description: |
        An entry in the append-only audit log. Every authorization decision,
        registration, key rotation, and admin mutation is recorded.
      required: [event_id, timestamp, event_type]
      properties:
        event_id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        event_type:
          type: string
          description: |
            Event type, e.g. `AUTHORIZE_CALLED`, `PROOF_ISSUED`,
            `DECISION_CREATED`, `AGENT_REGISTERED`, `OWNER_CREATED`,
            `POLICY_UPSERTED`, `OWNER_LOGIN`, `OWNER_LOGOUT`,
            `APPROVAL_REQUEST_CREATED`, `APPROVAL_REQUEST_APPROVED`,
            `APPROVAL_REQUEST_DENIED`, `APPROVAL_TOKEN_USED`,
            `SERVER_STARTED`.
        principal_id:
          type: string
          format: uuid
          nullable: true
        action_id:
          type: string
          format: uuid
          nullable: true
        decision_id:
          type: string
          format: uuid
          nullable: true
        metadata_json:
          type: object
          description: Event-specific details.
...